requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
python-dotenv==1.0.0

Office365-REST-Python-Client==2.5.3
openai==1.12.0
chromadb==0.4.22

PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

apscheduler==3.10.4
